Set up some proper inheritance
